Refactor EIP1559/EIP2930 transaction
